requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncio==3.4.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
currency-symbols==2.0.3
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
mysql-connector-python==9.0.0
pydantic==2.7.4
pydantic_core==2.18.4
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
